+++ twitch-follower-organizer.user.js
@@ -1,7 +1,7 @@
 // ==UserScript==
 // @name        Twitch Follower Organizer
 // @namespace   twitch-follower-organizer
-// @version     0.1.14
+// @version     0.1.15
 // @author      Nesswit
 // @description "We need better sidebar" - by wonzy_world, 2021
 // @supportURL  https://github.com/rishubil/twitch-follower-organizer/issues
@@ -27,12 +27,14 @@
   const CLIENT_ID = 'kimne78kx3ncx6brgo4mv6wki5h1ko';
   const UNKNOWN_GROUP_NAME = 'ETC';
   const GROUPS_VALUE_NAME = 'groups';
+  const GROUP_DEFUALT_COLOR = '#a970ff';
 
   /**
    * User defined channel group, with some states and options
    * @typedef {object} Group
    * @property {string} group_name - The name of group
    * @property {boolean} is_opened - The state of group is opened or not
+   * @property {boolean} is_locked - The state of group is locked or not
    * @property {boolean} hide_offline - Whether to show channels that are offline
    * @property {string} color - The Color of group
    * @property {?string[]} channels - List of channels in the group (UNKNOWN group has null value)
@@ -81,12 +83,18 @@
       {
         group_name: UNKNOWN_GROUP_NAME,
         is_opened: false,
+        is_locked: false,
         hide_offline: true,
-        color: '#a970ff',
+        color: GROUP_DEFUALT_COLOR,
         channels: null,
       },
     ];
     groups = GM_getValue(GROUPS_VALUE_NAME, default_groups);
+    _.forEach(groups, function (value) {
+      if (value['is_locked'] === undefined) {
+        value['is_locked'] = false;
+      }
+    });
   }
 
   /**
@@ -119,8 +127,9 @@
     groups.splice(0, 0, {
       group_name: group_name,
       is_opened: false,
+      is_locked: false,
       hide_offline: true,
-      color: '#a970ff',
+      color: GROUP_DEFUALT_COLOR,
       channels: [],
     });
     saveGroups();
@@ -132,11 +141,15 @@
    * If there is no group named `group_name`, do nothing.
    *
    * @param {string} group_name The name of group
+   * @throws If the group is locked
    */
   function removeGroup(group_name) {
     const group_index = findGroupIndexByName(group_name);
     if (findGroupIndexByName(group_name) === -1) {
       return;
+    }
+    if (groups[group_index]['is_locked']) {
+      throw new Error('LOCKED');
     }
     groups.splice(group_index, 1);
     saveGroups();
@@ -147,6 +160,7 @@
    * @param {number} source_group_index Source group index
    * @param {number} target_group_index Target group index
    * @param {string} channel_name Channel name to move
+   * @throws If the group is locked
    */
   function moveChannelBetweenGroups(
     source_group_index,
@@ -155,6 +169,12 @@
   ) {
     if (source_group_index === target_group_index) {
       return;
+    }
+    if (
+      groups[source_group_index]['is_locked'] ||
+      groups[target_group_index]['is_locked']
+    ) {
+      throw new Error('LOCKED');
     }
     const unknown_group_index = findGroupIndexByName(UNKNOWN_GROUP_NAME);
     if (source_group_index !== unknown_group_index) {
@@ -236,6 +256,7 @@
     if (group === null) {
       return;
     }
+    // It will be opened even if the group is locked
     group['is_opened'] = is_opened;
     saveGroups();
     renderFollowedSection();
@@ -663,7 +684,7 @@
         <div>
           <div class="side-nav-card tw-relative" data-test-selector="side-nav-card"><a
               class="tbs-group-item tbs-link side-nav-card__link tw-align-items-center tw-flex tw-flex-nowrap tw-full-width tw-link tw-link--hover-underline-none tw-pd-x-1 tw-pd-y-05"
-              data-test-selector="followed-channel" data-tbs-group-index="<%- group_index %>" data-tbs-channel="<%-  channel_info.user.login %>" href="/<%- channel_info.user.login %>" draggable="true">
+              data-test-selector="followed-channel" data-tbs-group-index="<%- group_index %>" data-tbs-channel="<%-  channel_info.user.login %>" href="/<%- channel_info.user.login %>" draggable="<%- draggable %>">
               <div class="side-nav-card__avatar <% if (!is_live) { %>side-nav-card__avatar--offline <% } %>tw-align-items-center tw-flex-shrink-0">
                 <figure aria-label="<%- channel_info.user.displayName %> (<%- channel_info.user.login %>)" class="tw-avatar tw-avatar--size-30"><img
                     class="tw-block tw-border-radius-rounded tw-image tw-image-avatar"
@@ -711,6 +732,7 @@
     `);
     return templateTbsGroupItem({
       group_index: group_index,
+      draggable: groups[group_index]['is_locked'] ? 'false' : 'true',
       channel_info: channel_info,
       is_live: channel_info.content.viewersCount !== undefined,
     });
@@ -861,6 +883,67 @@
   }
 
   /**
+   * get group from setting overlay
+   *
+   * @return {?Group} Return group if there is the group else null.
+   */
+  function getGroupFromSettingOverlray() {
+    let groupSettingEl = document.getElementsByClassName('tbs-group-setting');
+    if (groupSettingEl.length === 0) {
+      return null;
+    } else {
+      groupSettingEl = groupSettingEl[0];
+    }
+    const target_group_name = groupSettingEl.dataset.tbsGroupName;
+    const group = getGroupByName(target_group_name);
+    return group;
+  }
+
+  /**
+   * Reset group color input on setting overlay
+   *
+   * If there is no group setting overlay, do nothing.
+   * If the group is locked, do nothing.
+   */
+  function resetGroupColorOnSettingOverlay() {
+    let groupSettingEl = document.getElementsByClassName('tbs-group-setting');
+    if (groupSettingEl.length === 0) {
+      return;
+    } else {
+      groupSettingEl = groupSettingEl[0];
+    }
+    const target_group_name = groupSettingEl.dataset.tbsGroupName;
+    const group = getGroupByName(target_group_name);
+    if (group['is_locked']) {
+      return;
+    }
+    const colorEl = groupSettingEl.getElementsByClassName(
+      'tbs-group-setting-color'
+    )[0];
+    colorEl.value = GROUP_DEFUALT_COLOR;
+  }
+
+  /**
+   * Set group locked from setting overlay
+   *
+   * If there is no group setting overlay, do nothing.
+   *
+   * @return {?string} Return null if save successfully, else error text.
+   */
+  function setGroupLockedFromSettingOverlay(is_locked) {
+    const group = getGroupFromSettingOverlray();
+    if (group['group_name'] === UNKNOWN_GROUP_NAME) {
+      return '잠금 여부를 변경할 수 없는 그룹입니다.';
+    }
+
+    group['is_locked'] = is_locked;
+
+    saveGroups();
+
+    return null;
+  }
+
+  /**
    * Save group from setting overlay
    *
    * If there is no group setting overlay, do nothing.
@@ -876,6 +959,9 @@
     }
     const target_group_name = groupSettingEl.dataset.tbsGroupName;
     const group = getGroupByName(target_group_name);
+    if (group['is_locked']) {
+      return '그룹이 잠겨있습니다.';
+    }
 
     const group_name = groupSettingEl.getElementsByClassName(
       'tbs-group-setting-group-name'
@@ -921,13 +1007,12 @@
    * @return {?string} Return null if delete successfully, else error text.
    */
   function deleteGroupFromSettingOverlay() {
-    let groupSettingEl = document.getElementsByClassName('tbs-group-setting');
-    if (groupSettingEl.length === 0) {
-      return;
-    } else {
-      groupSettingEl = groupSettingEl[0];
+    const group = getGroupFromSettingOverlray();
+    const target_group_name = group['group_name'];
+
+    if (group['is_locked']) {
+      return '그룹이 잠겨있습니다.';
     }
-    const target_group_name = groupSettingEl.dataset.tbsGroupName;
 
     if (target_group_name === UNKNOWN_GROUP_NAME) {
       // UNKNOWN group cannot be deleted
@@ -982,23 +1067,42 @@
                     <input type="text"
                       class="tbs-group-setting-group-name tw-block tw-border-bottom-left-radius-medium tw-border-bottom-right-radius-medium tw-border-top-left-radius-medium tw-border-top-right-radius-medium tw-font-size-6 tw-full-width tw-input tw-pd-l-1 tw-pd-r-1 tw-pd-y-05"
                       autocapitalize="off" autocorrect="off" autocomplete="off"
-                      spellcheck="false" value="<%- group['group_name'] %>"<% if (is_unknown_group) { %> disabled="disabled"<% } %>>
+                      spellcheck="false" value="<%- group['group_name'] %>"<% if (is_unknown_group || group['is_locked']) { %> disabled="disabled"<% } %>>
                   </div>
                   <div>
                     <div class="tw-mg-b-05 tw-mg-t-1">
                       <label class="tw-form-label">색상</label>
                     </div>
-                    <input type="text"
-                      class="tbs-group-setting-color tw-block tw-border-bottom-left-radius-medium tw-border-bottom-right-radius-medium tw-border-top-left-radius-medium tw-border-top-right-radius-medium tw-font-size-6 tw-full-width tw-input tw-pd-l-1 tw-pd-r-1 tw-pd-y-05"
-                      autocapitalize="off" autocorrect="off" autocomplete="off"
-                      spellcheck="false" value="<%- group['color'] %>">
+                    <div class="tw-combo-input tw-flex tw-full-width">
+                      <div class="tw-combo-input__input tw-flex-grow-1">
+                        <div class="tw-relative">
+                          <input type="text" class="tbs-group-setting-color tw-block tw-border-bottom-left-radius-medium tw-border-bottom-right-radius-none tw-border-top-left-radius-medium tw-border-top-right-radius-none tw-font-size-6 tw-full-width tw-input tw-pd-l-1 tw-pd-r-1 tw-pd-y-05" autocapitalize="off" autocorrect="off" autocomplete="off"
+                          spellcheck="false" value="<%- group['color'] %>"<% if (group['is_locked']) { %> disabled="disabled"<% } %>>
+                        </div>
+                      </div>
+                      <button aria-label="색상 초기화" class="tbs-group-setting-color-reset-button tw-align-items-center tw-align-middle tw-border-bottom-left-radius-none tw-border-bottom-right-radius-medium tw-border-top-left-radius-none tw-border-top-right-radius-medium tw-combo-input__button-icon tw-core-button tw-core-button--secondary tw-inline-flex tw-justify-content-center tw-overflow-hidden tw-relative<% if (group['is_locked']) { %> tw-core-button--disabled<% } %>">
+                        <div class="tw-align-items-center tw-core-button-icon tw-inline-flex">
+                          <div style="width: 2rem; height: 2rem;">
+                            <div class="tw-icon">
+                              <div class="tw-aspect">
+                                <svg xmlns="http://www.w3.org/2000/svg" class="tbs-button-svg" width="100%" height="100%" viewBox="0 0 24 24" stroke-width="2" fill="none" stroke-linecap="round" stroke-linejoin="round">
+                                  <path stroke="none" d="M0 0h24v24H0z" fill="none"/>
+                                  <path d="M20 11a8.1 8.1 0 0 0 -15.5 -2m-.5 -4v4h4" />
+                                  <path d="M4 13a8.1 8.1 0 0 0 15.5 2m.5 4v-4h-4" />
+                                </svg>
+                              </div>
+                            </div>
+                          </div>
+                        </div>
+                      </button>
+                    </div>
                   </div>
                   <div>
                     <div class="tw-mg-b-05 tw-mg-t-1">
                       <label class="tw-form-label">오프라인 채널 표시 여부</label>
                     </div>
                     <select
-                      class="tbs-group-setting-hide-offline tw-block tw-border-radius-medium tw-font-size-6 tw-full-width tw-pd-l-1 tw-pd-r-3 tw-pd-y-05 tw-select">
+                      class="tbs-group-setting-hide-offline tw-block tw-border-radius-medium tw-font-size-6 tw-full-width tw-pd-l-1 tw-pd-r-3 tw-pd-y-05 tw-select"<% if (group['is_locked']) { %> disabled="disabled"<% } %>>
                       <option value="false"<% if (!group['hide_offline']) { %>selected="selected"<% } %>>표시함</option>
                       <option value="true"<% if (group['hide_offline']) { %>selected="selected"<% } %>>표시하지 않음</option>
                     </select>
@@ -1006,24 +1110,61 @@
                 </div>
                 <div class="tw-mg-t-1 tw-align-items-center tw-flex tw-flex-grow-1 tw-flex-shrink-1 tw-full-width tw-justify-content-between">
                   <% if (!is_unknown_group) { %>
-                    <button aria-label="삭제"
-                    class="tbs-group-setting-delete-button tw-align-items-center tw-align-middle tw-border-bottom-left-radius-medium tw-border-bottom-right-radius-medium tw-border-top-left-radius-medium tw-border-top-right-radius-medium tw-button-icon tw-core-button tw-inline-flex tw-justify-content-center tw-overflow-hidden tw-relative"><span
-                      class="tw-button-icon__icon">
-                      <div style="width: 2rem; height: 2rem;">
-                        <div class="tw-icon">
-                          <div class="tw-aspect">
-                            <svg xmlns="http://www.w3.org/2000/svg" class="tbs-button-svg" width="100%" height="100%" viewBox="0 0 24 24"stroke-width="2" fill="none" stroke-linecap="round" stroke-linejoin="round">
-                              <path stroke="none" d="M0 0h24v24H0z" fill="none"/>
-                              <line x1="4" y1="7" x2="20" y2="7" />
-                              <line x1="10" y1="11" x2="10" y2="17" />
-                              <line x1="14" y1="11" x2="14" y2="17" />
-                              <path d="M5 7l1 12a2 2 0 0 0 2 2h8a2 2 0 0 0 2 -2l1 -12" />
-                              <path d="M9 7v-3a1 1 0 0 1 1 -1h4a1 1 0 0 1 1 1v3" />
-                            </svg>
+                    <div>
+                      <button aria-label="삭제"
+                      class="tbs-group-setting-delete-button tw-align-items-center tw-align-middle tw-border-bottom-left-radius-medium tw-border-bottom-right-radius-medium tw-border-top-left-radius-medium tw-border-top-right-radius-medium tw-button-icon tw-core-button tw-inline-flex tw-justify-content-center tw-overflow-hidden tw-relative<% if (group['is_locked']) { %> tw-core-button--disabled<% } %>"><span
+                        class="tw-button-icon__icon">
+                        <div style="width: 2rem; height: 2rem;">
+                          <div class="tw-icon">
+                            <div class="tw-aspect">
+                              <svg xmlns="http://www.w3.org/2000/svg" class="tbs-button-svg" width="100%" height="100%" viewBox="0 0 24 24"stroke-width="2" fill="none" stroke-linecap="round" stroke-linejoin="round">
+                                <path stroke="none" d="M0 0h24v24H0z" fill="none"/>
+                                <line x1="4" y1="7" x2="20" y2="7" />
+                                <line x1="10" y1="11" x2="10" y2="17" />
+                                <line x1="14" y1="11" x2="14" y2="17" />
+                                <path d="M5 7l1 12a2 2 0 0 0 2 2h8a2 2 0 0 0 2 -2l1 -12" />
+                                <path d="M9 7v-3a1 1 0 0 1 1 -1h4a1 1 0 0 1 1 1v3" />
+                              </svg>
+                            </div>
                           </div>
                         </div>
-                      </div>
-                    </span></button>
+                      </span></button>
+                      <% if (group['is_locked']) { %>
+                        <button aria-label="잠금 해제"
+                        class="tbs-group-setting-unlock-button tw-align-items-center tw-align-middle tw-border-bottom-left-radius-medium tw-border-bottom-right-radius-medium tw-border-top-left-radius-medium tw-border-top-right-radius-medium tw-button-icon tw-core-button tw-inline-flex tw-justify-content-center tw-overflow-hidden tw-relative"><span
+                          class="tw-button-icon__icon">
+                          <div style="width: 2rem; height: 2rem;">
+                            <div class="tw-icon">
+                              <div class="tw-aspect">
+                                <svg xmlns="http://www.w3.org/2000/svg" class="tbs-button-svg" width="100%" height="100%" viewBox="0 0 24 24"stroke-width="2" fill="none" stroke-linecap="round" stroke-linejoin="round">
+                                  <path stroke="none" d="M0 0h24v24H0z" fill="none"/>
+                                  <rect x="5" y="11" width="14" height="10" rx="2" />
+                                  <circle cx="12" cy="16" r="1" />
+                                  <path d="M8 11v-5a4 4 0 0 1 8 0" />
+                                </svg>
+                              </div>
+                            </div>
+                          </div>
+                        </span></button>
+                      <% } else { %>
+                        <button aria-label="잠금"
+                        class="tbs-group-setting-lock-button tw-align-items-center tw-align-middle tw-border-bottom-left-radius-medium tw-border-bottom-right-radius-medium tw-border-top-left-radius-medium tw-border-top-right-radius-medium tw-button-icon tw-core-button tw-inline-flex tw-justify-content-center tw-overflow-hidden tw-relative"><span
+                          class="tw-button-icon__icon">
+                          <div style="width: 2rem; height: 2rem;">
+                            <div class="tw-icon">
+                              <div class="tw-aspect">
+                                <svg xmlns="http://www.w3.org/2000/svg" class="tbs-button-svg" width="100%" height="100%" viewBox="0 0 24 24"stroke-width="2" fill="none" stroke-linecap="round" stroke-linejoin="round">
+                                  <path stroke="none" d="M0 0h24v24H0z" fill="none"/>
+                                  <rect x="5" y="11" width="14" height="10" rx="2" />
+                                  <circle cx="12" cy="16" r="1" />
+                                  <path d="M8 11v-4a4 4 0 0 1 8 0v4" />
+                                </svg>
+                              </div>
+                            </div>
+                          </div>
+                        </span></button>
+                      <% } %>
+                    </div>
                   <% } else { %>
                     <div></div>
                   <% } %>
@@ -1044,7 +1185,7 @@
                         </div>
                       </span></button>
                     <button aria-label="저장"
-                      class="tbs-group-setting-save-button tw-align-items-center tw-align-middle tw-border-bottom-left-radius-medium tw-border-bottom-right-radius-medium tw-border-top-left-radius-medium tw-border-top-right-radius-medium tw-button-icon tw-core-button tw-inline-flex tw-justify-content-center tw-overflow-hidden tw-relative"><span
+                      class="tbs-group-setting-save-button tw-align-items-center tw-align-middle tw-border-bottom-left-radius-medium tw-border-bottom-right-radius-medium tw-border-top-left-radius-medium tw-border-top-right-radius-medium tw-button-icon tw-core-button tw-inline-flex tw-justify-content-center tw-overflow-hidden tw-relative<% if (group['is_locked']) { %> tw-core-button--disabled<% } %>"><span
                         class="tw-button-icon__icon">
                         <div style="width: 2rem; height: 2rem;">
                           <div class="tw-icon">
@@ -1080,7 +1221,7 @@
    * @param {string} class_name class name to find
    * @param {number} max_bubble_count maximum recursive depth, default is 10
    */
-  function findEventTargetbyClassName(
+  function findEventTargetByClassName(
     event,
     class_name,
     max_bubble_count = 20
@@ -1121,7 +1262,7 @@
       'click',
       function (e) {
         if (e.target) {
-          const addGroupButton = findEventTargetbyClassName(
+          const addGroupButton = findEventTargetByClassName(
             e,
             'tbs-add-group-button'
           );
@@ -1146,7 +1287,7 @@
               }
             }
           }
-          const editButton = findEventTargetbyClassName(e, 'tbs-edit-button');
+          const editButton = findEventTargetByClassName(e, 'tbs-edit-button');
           if (editButton !== null) {
             const group_index = Number(editButton.dataset.tbsGroupIndex);
             const group = groups[group_index];
@@ -1166,7 +1307,60 @@
             e.preventDefault();
             return;
           }
-          const groupSettingCancelButton = findEventTargetbyClassName(
+          const groupSettingColorResetButton = findEventTargetByClassName(
+            e,
+            'tbs-group-setting-color-reset-button'
+          );
+          if (groupSettingColorResetButton !== null) {
+            const group = getGroupFromSettingOverlray();
+            if (group !== null && group['is_locked']) {
+              return;
+            }
+            resetGroupColorOnSettingOverlay();
+            e.preventDefault();
+            return;
+          }
+          const groupSettingLockButton = findEventTargetByClassName(
+            e,
+            'tbs-group-setting-lock-button'
+          );
+          if (groupSettingLockButton !== null) {
+            const group = getGroupFromSettingOverlray();
+            const setResult = setGroupLockedFromSettingOverlay(true);
+            if (setResult === null) {
+              showGroupSettingOverlay(group);
+              processFollowedSectionData();
+            } else {
+              const errorEl = document.getElementsByClassName(
+                'tbs-group-settings-error'
+              )[0];
+              errorEl.style.display = 'block';
+              errorEl.innerHTML = setResult;
+            }
+            e.preventDefault();
+            return;
+          }
+          const groupSettingUnlockButton = findEventTargetByClassName(
+            e,
+            'tbs-group-setting-unlock-button'
+          );
+          if (groupSettingUnlockButton !== null) {
+            const group = getGroupFromSettingOverlray();
+            const setResult = setGroupLockedFromSettingOverlay(false);
+            if (setResult === null) {
+              showGroupSettingOverlay(group);
+              processFollowedSectionData();
+            } else {
+              const errorEl = document.getElementsByClassName(
+                'tbs-group-settings-error'
+              )[0];
+              errorEl.style.display = 'block';
+              errorEl.innerHTML = setResult;
+            }
+            e.preventDefault();
+            return;
+          }
+          const groupSettingCancelButton = findEventTargetByClassName(
             e,
             'tbs-group-setting-cancel-button'
           );
@@ -1175,11 +1369,15 @@
             e.preventDefault();
             return;
           }
-          const groupSettingSaveButton = findEventTargetbyClassName(
+          const groupSettingSaveButton = findEventTargetByClassName(
             e,
             'tbs-group-setting-save-button'
           );
           if (groupSettingSaveButton !== null) {
+            const group = getGroupFromSettingOverlray();
+            if (group !== null && group['is_locked']) {
+              return;
+            }
             const saveResult = saveGroupFromSettingOverlay();
             if (saveResult === null) {
               clearGroupSettingOverlay();
@@ -1194,11 +1392,15 @@
             e.preventDefault();
             return;
           }
-          const groupSettingDeleteButton = findEventTargetbyClassName(
+          const groupSettingDeleteButton = findEventTargetByClassName(
             e,
             'tbs-group-setting-delete-button'
           );
           if (groupSettingDeleteButton !== null) {
+            const group = getGroupFromSettingOverlray();
+            if (group !== null && group['is_locked']) {
+              return;
+            }
             const deleteResult = deleteGroupFromSettingOverlay();
             if (deleteResult === null) {
               clearGroupSettingOverlay();
@@ -1213,6 +1415,7 @@
             e.preventDefault();
             return;
           }
+
           const sideNavToggleButton = findEventTargetbyClassName(
             e,
             'collapse-toggle'
@@ -1235,7 +1438,7 @@
               return;
             }
           }
-          const link = findEventTargetbyClassName(e, 'tbs-link');
+          const link = findEventTargetByClassName(e, 'tbs-link');
           if (link !== null) {
             if (
               e.getModifierState('Alt') ||
@@ -1253,7 +1456,7 @@
             e.preventDefault();
             return;
           }
-          const group_setting = findEventTargetbyClassName(
+          const group_setting = findEventTargetByClassName(
             e,
             'tbs-group-setting'
           );
@@ -1271,7 +1474,7 @@
           return;
         }
         if (e.target) {
-          const card = findEventTargetbyClassName(e, 'side-nav-card__link');
+          const card = findEventTargetByClassName(e, 'side-nav-card__link');
           if (card !== null) {
             if (card.classList.contains('tbs-group-item')) {
               const group_index = Number(card.dataset.tbsGroupIndex);
@@ -1289,7 +1492,7 @@
             }
           }
 
-          const cardOverlay = findEventTargetbyClassName(e, 'tbs-card-overlay');
+          const cardOverlay = findEventTargetByClassName(e, 'tbs-card-overlay');
           if (cardOverlay !== null) {
             shouldShowCardOverlay = true;
             e.preventDefault();
@@ -1303,7 +1506,7 @@
       'mouseout',
       function (e) {
         if (e.target) {
-          const card = findEventTargetbyClassName(e, 'side-nav-card__link');
+          const card = findEventTargetByClassName(e, 'side-nav-card__link');
           if (card !== null) {
             if (card.classList.contains('tbs-group-item')) {
               shouldShowCardOverlay = false;
@@ -1313,7 +1516,7 @@
             }
           }
 
-          const cardOverlay = findEventTargetbyClassName(e, 'tbs-card-overlay');
+          const cardOverlay = findEventTargetByClassName(e, 'tbs-card-overlay');
           if (cardOverlay !== null) {
             shouldShowCardOverlay = false;
             debouncedClearCardOverlay();
@@ -1328,7 +1531,7 @@
       'dragstart',
       function (e) {
         if (e.target) {
-          const card = findEventTargetbyClassName(e, 'side-nav-card__link');
+          const card = findEventTargetByClassName(e, 'side-nav-card__link');
           if (card !== null) {
             shouldShowCardOverlay = false;
             debouncedClearCardOverlay();
@@ -1378,15 +1581,35 @@
       'drop',
       function (e) {
         if (e.target) {
-          const card = findEventTargetbyClassName(e, 'side-nav-card__link');
+          const card = findEventTargetByClassName(e, 'side-nav-card__link');
           if (card !== null && dragged_card !== null) {
             const dragged_group_index = Number(
               dragged_card.dataset.tbsGroupIndex
             );
             const group_index = Number(card.dataset.tbsGroupIndex);
             if (dragged_card.classList.contains('tbs-group-header')) {
-              moveGroupPosition(dragged_group_index, group_index);
+              const group_rect = card
+                .closest('.tbs-group')
+                .getBoundingClientRect();
+              if (dragged_group_index === group_index) {
+                // do nothing
+              } else if (dragged_group_index < group_index) {
+                if (e.clientY < group_rect.top + group_rect.height / 2) {
+                  moveGroupPosition(dragged_group_index, group_index - 1);
+                } else {
+                  moveGroupPosition(dragged_group_index, group_index);
+                }
+              } else {
+                if (e.clientY < group_rect.top + group_rect.height / 2) {
+                  moveGroupPosition(dragged_group_index, group_index);
+                } else {
+                  moveGroupPosition(dragged_group_index, group_index + 1);
+                }
+              }
             } else if (dragged_card.classList.contains('tbs-group-item')) {
+              if (groups[group_index]['is_locked']) {
+                return;
+              }
               const dragged_channel_name = dragged_card.dataset.tbsChannel;
               moveChannelBetweenGroups(
                 dragged_group_index,